Create an example AI project #38
Example project is classifying animals as either reptiles or mammals using a
decision tree classifier based on 3 attributes about the animal.

The example project has been added to test the functionality of the base application
and so that fellow developers can use it to base their own machine learning
projects from it.